+++ CHANGES.md
@@ -1,3 +1,4 @@
+ - Throw if callback supplied to "finally" is invalid (@grahamrhay)
 
 ## 1.4.1